Establish link for existing entities
Previously we were not establishing a link between an entity and an existing
entity; this was to avoid the scenario whereby, for example, a candidate is
created with a qualification and later updated. I was under the impression
re-linking an existing entity (in this case the existing qualification) would
result in a duplicate relationship. This is not the case and it turns out its
safe to re-link existing entities.

It also has the benefit of being able to link new entities to existing
entities. This is requried for the work to create teaching events; some will be
associated with existing buildings and with the code as it was before the link
was not being established.