Fix TeachingEvent not linking to Buildings
The teaching events were not being linked to the correct building as the
`TeachingEvent.BuildingId` field was not being mapped from the CRM entity (this
is then used to establish the relationship before persisting the event).

Update to map the `BuildingId` so it gets populated correctly and the building
gets linked.