Tweak ReadableIdHasChanged logic
Previously we were checking for a `null` teaching event `Id` in the wider
`When` conditional and within the method itself, which results in a block of
dead code (as the `null` id check within the method could never be reached).

Also adds test for the scenario where the teaching event has an `Id` but it
can't be found in the store.