Remove CrmETag
A new ETag is generated for successful CRM syncs. This was implemented to prevent multiple requests for unchanged resources.

However, this was implemented at a time when the CRM sync ran every hour. Since then, it was requested that the CRM sync runs every 5 minutes instead.

This means that the ETag is no longer useful. Cache-Control is used as well (also with an expiration time of 5 minutes) that will ensure that requests are not made anyway.