Improve logging around potential duplicate Candidates
If the CRM is responding it should not be possible for a duplicate candidate to
be created in the system (such that the email address, first name and last name
match an existing candidate).

We need to improve the logging around the scenarios when the CRM is not
responding, which should either be when the integration is paused (i.e. they
are updating the CRM) or an unexpected exception occurs (the request times
out/the CRM is down/etc).

Add improved logging around matchback/access token generation.