Support compact logging of verbose requests/responses
Currently we enable request/response logging for all API endpoints, however
some of these endpoints return quite a lot of not necessarily all that useful
information when it comes to debugging issues. Its mainly the POST/PUT
request/response payloads that we want to log out (but not exclusively - there
are some GET request payloads like the healthcheck that are useful).

Add support for configuring the `RequestResponseLoggingMiddleware` with a set
of `CompactLoggingPattern`s - any request/response that matches a comapct regex
pattern will not have its payload included in the log.