In-memory cache postcodes that can't be geocoded
If a teaching event is assigned a postcode that passes validation but cannot be
looked up successfuly from either the local database of postcode geolocations
or via the Google Geocoding API then we can end up re-querying the postcode
multiple times each sync. This can lead to very noisy alerts going off as we
have a low-tolerance for Google Geocoding API failures (a single failure is
notified).

Instead, we should cache the postcodes that fail lookup so that we don't
continually receive alerts about them and we can then avoid calling the Google
Geocoding API for a known bad postcode.