Add modelling for classroom experience
A `Candidate` is able to have associated classroom experience inotes with their
`contact` record. This is stored as a single field in the CRM on the `contact`
entity, however we abstract this from the clients with the
`ClassroomExperienceNote` model. It accepts the required parameters and has a
`ToString()` method to format it appropriately for the underlying field.

Add string extension to nullify an empty string; we make use of this in
`Candidate.FullName` to avoid it being included in the change tracking when not
set (previously it was an empty string which would count as a change on the
computer property).