Call `DeleteLink` on `TeachingEvent` when `Building` is removed
Existing relationships between `TeachingEvent` and `TeachingEventBuilding` do not get removed when the related Building is deleted. Currently, the Event is persisted in the cache with a null Building, which will then be re-added when the CRM sync runs.

Call `DeleteLink` in this scenario to remove the relationship in the CRM as well.