Update SE upsert operation to be synchronous
Instead of queueing the job to upsert a candidate we need to be able to return
the newly created candidate (with an ID) to the Schools Experience service.

This means the service will be unavailable when the CRM is offline, but that is
the case currenly anyway.

Once the migration to the API has been completed we may be able to udpate the
service to remove this requirement.