Add support for comma-separated query string values
By default ASP.NET Core expects `FromQuery` parameters of a collectiont type to
be expressed in the format:

```
GET /items?key=value1&key=value2
```

Swagger UI and SwaggerCodegen, however, create code that results in the query
parameters in the format:

```
GET /items?key=value1,value2
```

As it doesn't appear possible to customise Swagger UI/SwaggerCodegen in this
respect, this commit instead updates ASP.NET Core with a `CommaSeparated`
attribute that can be applied to controller parameters in order to accept the
comma-separated value format.

It does this via a mixture of custom attributs, value providers and a
convention - all of which are closed to mocking and difficult to test, so I've
covered the bits of code that I can.