Make logical name explicit for all EntityReference attributes
The old `CdsServiceClient` was able to infer the entity logical name for
relationships, however the updated `ServiceClient` doesn't appear to do this
and requires the logical name to be explicitly provided. It was already
provided in most cases, but a couple of inferred attributes on the
`TeachingEventRegistration` was causing issues.