Strip exit code before serving existing sign ups
If a candidate already exists in the system its likely that we will have
transformed their telephone number into a format that is dialiable by TPUK's
telephone system. This means it will be prefixed with a `00` exit code. As we
don't want the user to have to put in the exit code manually we expect their
country dial-in code at the start of the telephone instead. We need to
therefore strip the `00` exit code from any responses that are subsequently
used to pre-fill a form.

Reworks the telephone sanitization to pull it into a string extension so that
we can more easily reuse it elsewhere.