Add generic UpsertModelJob
As we are splitting the `UpsertCandidateJob` into multiple independent jobs we
need to be able to upsert a variety of records that have a `CandidateId` (are
related to a candidate). This commit adds an `UpsertModelJob` that can take any
`BaseModel` implementing `IHasCandidateId` and upsert it to the CRM.

Update applicable models to implement the new interface.