Have CandidateUpserter queue jobs for sub-tasks
The main purpose of the `CandidateUpserter` is to upsert a `Candidate` record.
As a by-product it also supports upserting all the objects related to a
candidate, such as accepted privacy policy, qualifications, phone calls etc.

Currently this is all done in the `CandidateUpserter`, however as this is all
wrapped in a single job if any part fails and it retries it could cause a
duplicate record to be created.

To get around this, the `CandidateUpserter` is updated to queue more jobs for
each upsert sub-task (once the candidate is successfully upserted). This way,
if anything fails its retried independently and duplicates should be avoided.