On match back failure log error to Sentry
If the CRM is down/throws any kind of exception we currently swallow it and
return 404, so the user can continue as a new candidate. Instead, we want to
raise the error to Sentry so we don't loose sight of it (if there is a bug in
the matchback query, for example).