Support match back on email address only
The normal match back process involves matching the email plus 2/3 of first
name, last name and date of birth.

As the Find & Apply API is only going to be exposing the email address of the
candidates initially we need to support matching back on just that field.

Replicates the current match back logic (taking into account status, duplicate
score and date modified) and picks the first with the correct email.