Run MagicLinkTokenGenerationJob hourly
We run this minutely at present, but the feature is not in use - to avoid it
polluting the logs and to make debugging other jobs easier I'm switching it to
run hourly; if we need it in the future we can shorten the interval again. I
don't want to turn it off completely as having it ticking over would surface
any bugs/errors in the code.