Add basic auth to Hangfire dashboard
The Hangfire dashboard is only available in the dev/test environment, however
there is no authentication in front of it at the moment. Whilst its not
particularly exploitable, someone could requeue a bunch of jobs and potentially
clog up our workers.

Add basic authentication in front of the Hangfire dashboard, enabled only in
non-development environments.