Upgrade to .NET Core 5/all dependencies
Upgrade the two project files to .NET Core 5 - whilst this version does not
have LTS it will be supported until the end of next year, by which point we
should have already upgraded to the next version with LTS; upgrading now should
make that process easier (so we're not jumping from 3 all the way to 6).

Bring all dependencies up to the latest available versions.