Sync teaching events up to 9 months old
We were finding that the app was serving 404s for some past teaching events;
users likely bookmark the page and revisit it once the 4 month TTL of the
events cache has expired. As the event sync was optimised a while back we
should be fine to increase this to retrieve events further into the past.

Sync events going back 9 months instead of 4.