Sync Question Time events as Train to Teach
We have an immediate need to display Question Time events on the GiT website,
bundled in with Train to Teach events. Doing this on the front-end is a
substantial change that may not be doable by the deadline of tomorrow, so for
now we are going to re-type these events during the API sync so that the
website will automatically pull them through and display them correctly.

Once this is in we will look to correctly handle QT events on the website
directly.