Check if candidate can re-subcsribe based on adviser status
Up until now if a candidate has already been assigned an adviser we prevent
them from re-signing up to the TTA service.

It turns out some candidates that have been assigned an adviser will have their
`contact` record set to have a 'closed' adviser status. In this scenario, we
want to let them re-subscribe for an adviser.

Adds a new `CanSubscribeToTeacherTrainingAdviser` attribute as
`AlreadySubscribedToTeacherTrainingAdviser` would not be completely accurate
for the new use case. Once the TTA service has been updated this field can be
removed.