Don't swallow CRM exceptions on token generation
We were swallowing CRM exceptions on the matchback endpoint and, if
encountered, letting the candidate continue as a new `contact` record (logging
and raising to Sentry).

This is good from a user point of view, but creates an issue during development
whereby if a schema change isn't present in one of the CRM environments and the
error is swallowed its difficult to pin down what the problem is.

Instead, we're going to re-raise the error -- as the CRM team pause the
integration when the CRM is offline an error should never happen and if it does
its probably something wrong that we need to go in and fix.