Add IsWalkIn attribute to TeachingEventAddAttendee
We are going to start accepting 'walk in' sign ups to events that are otherwise
closed to new registrations.

Add `IsWalkIn` attribute to the `TeachingEventAddAttendee` model so we can
capture this and submit it with the rest of the event sign up form in the GiT
app.

When the new channel creation has been added to the CRM we will be able to mark
the candidate as a 'walk in' sign up.