Disable HTTP -> HTTPS redirection
A HTTP to HTTPS redirection is potentially insecure; on a browser-facing
website its often the lesser of two evils and forces users to connect via HTTPS
if they input the HTTP address. On an API this isn't necessary and we can be
more secure by disabling HTTP access completely, only accepting direct HTTPS
connections (preventing a potential man in the middle attack).

To do this we ensure the applicationn only acceps requests that are over HTTPS, as recommended by:

https://docs.microsoft.com/en-us/aspnet/core/security/enforcing-ssl?view=aspnetcore-5.0&tabs=visual-studio