Improve wait logic around contract test job queue
The contract tests queue multiple jobs, which queue other jobs themselves. I
think this process may be asynchronously happening in Hangfire and it results
in the queue being 'cleared' for short periods between one job finishing and
the next being enqueued.

To avoid flakiness we now perform a repeated check on the job queue when its in
the 'cleared' state (no in-progress/queued jobs) to ensure it stays in that
state for a short duration, which will hopefully resolve the flakiness in the
contract tests.