Switch EntityFieldAttribute Ignored attribute to use features
We had a `Ignored` attribute that worked off the environment name. This works
for most scenarios, but has the limitation of not being able to differentiate
between our hosted dev and test environments, as they both run with the ASP.NET
Core 'Staging' environment.

By switching it over to feature flags we can control which of these map to the
CRM more granularly, so if the CRM team have only deployed a schema change to
dev and not test, we can still deploy all the way to production and enable the
specific fields only in the dev environment.