Allow multiple event types in event search
Sometimes we want to search by multiple event types so that we don't have to filter events in the client. Add a TypeIds property that accepts an array of type IDs.

For now we are keeping the TypeId property for backwards compatibility which will be removed when the clients have been updated. Update any use of TypeId to TypeIds throughout the project, and temporarily set TypeIds to the value of TypeId if it is present.