Map phase/status to FindApply application forms
Previously we were mapping the phase/status at the candidate-level. Instead,
the Apply team have udpated their API to provide the phase/status for each
application form and will be deprecating the fields at the candidate-level.

The CRM team want to retain the candidate-level phase/status, having it contain
the latest applicable application forms details; this is inferred from the
first application form in the list we get back.

This commit maps the additional fields on `ApplicationForm` from the apply and
CRM models, infers the candidate-level values and updates the store to sync the
additional option sets.