Update with client configuration recommendations
A penetration test highlighted our insecure http redirect as a possible point of exploitation. Looking into disabling http traffic it turned out that given the current infrastructure its not currently possible without a substantial change of how the API is hosted (as the redirect happens at the infrastructure level).

Instead, the readme is being updated with client configuration recommendations to ensure all clients point directly at the `http` version of the API.