Make first/last name optional on Candidate
The candidates that come through from the Apply API will not always have a
`FirstName` or `LastName`. We need to be able to accept records where a name
has not been set, but still enforce first/last name on sign ups. As the request
models already have validation in place to ensure first/last name is present,
we can simply remove the `NotEmpty` check from the base `Candidate` model.