Fallback to matching on email only
We are now accepting candidates from Apply that only have an email address and
no first/last name, however we need to be able to match back on these candidate
records. In order to do this, we first attempt a 'full' match back on
first/last/email and, if that fails, we perform a second match back lookup
using only the email address.