Sort entities and attributes prior to comparison
Looking at the diff when the contract test fails in CI it appears to be because
the ordering of the entity attributes has changed. As the attributes are not a
collection the ordering is probably undefined and we can't sort it prior to
serializing. Instead, we deserialize the snapshot and then sort it manually
before performing the comaprison.