Allow ChannelId to be specified on TeacherTrainingAdviserSignUp
The events team are pushing the TTA sign up at face-to-face events and want to
be able to attribute sign ups back to the event.

Expose the `ChannelId` in the request model and map it to the `Candidate` (if a
new candidate) and `TeacherTrainingAdviserSignUp` on sign up.

The field is optional so that existing sign ups still work as expected and the
field is validated by the nested `CandidateValidator`.