Add `daily` label to Google CSP violations alert
This alert was going off continuously during the hours defined in the rule expression. A `daily` route has been added to the alertmanager config (https://github.com/DFE-Digital/cf-monitoring/pull/34) which means that this shouldn't happen anymore. 

Also, add to the expression that the alert should only go off on weekdays.
sd