Allow exchanging unverified request for event sign up
We are allowing walk-in event registrations to continue as an existing user
without verifying their access token. This is in case they don't have access to
the inbox or internet connectivity at the venue is poor and they don't receive
the code.

We will be flagging the contact record in the CRM as 'unverified' when this
mechanism is used to keep track of those that have chosen to bypass the
verification proceedure. We will also not be exposing pre-existing personal
information.