Add operations endpoint to run Find/Apply backfill
We don't want to run the backfill job on a schedule and instead want to be able
to trigger it manually. To achieve this a new endpoint has been added to the
`OperationsController` that will let us trigger a backfill job as long as one
is not already in progress.