Schedule backfill jobs in the future
When running the backfill in the pre-prod CRM environment using the prod Apply
API we found that the backfill job was eventually timing out due to the jobs it
spawned overloading the CRM. Instead, we want the backfill job to complete
before the individual candidate sync jobs begin to run. To achieve this we
schedule them for 30 mintues in the future (the backfill job takes around 20
minutes to complete when ran locally against production).

Increase the lock duration of the job to 60 minutes to ensure a second job can
never be enqueued while the first one is running.