Enable filtering by online/in-person events
The new event design will let users specify if they wish to attend in-person vs
online events; the API has been updated to support passing an `online` param in
the search request. If set, it will return only online events (if `true`) or
only in-person events (if `false`). If not specified the existing behaviour is
retained, returning both online and in-person events.