Add `Candidate` cache
The CRM can't accept any new data during releases. The solution to this in Get into Teaching is to queue the request until the CRM is back up.

However, in School Experience, the Candidate upsert cannot be queued because the candidate information is needed immediately by the Schools Experience app.

Add a `Candidate` cache and a intermediate table. The cache can be used when the CRM integration is paused. We can persist a temporary ID in the intermediate table which can be used to retrieve the CRM Candidate record once the cache has been cleared.