Add endpoint to get candidate by intermediate ID
If a candidate is created when the CRM is down, we can store the candidate intermediate ID in the School Experience app.

When we want to retrieve the record, we can check whether it still exists in the cache. If it does, return it, else lookup the CRM ID from the mapping table (this will populate the CRM ID for any further requests)