Correctly handle when Redis key is not yet set
I was expecting Redis to return `null` when a key is not yet set, however it
'returns a special nil value' which appears to be `nil` in a string. When we
pass this to the date parser it spits out a very early date, having the
knock-on effect of causing the normal Apply API sync job to essentially perform
a backfill.

To fix this we check if the key exists and manually return `null` or `false`
depending on the app settings method.