Sort entities before writing contract snapshot
We currently sort entities after reading the snapshot JSON, which means if the
snapshots change the diff can be quite large and hard to read (as the order its
saved in doesn't currently matter/is not consistent).

By sorting the entities prior to saving them it should make the diff
smaller/more readable.