Add relationship to application forms
Currently we persist application forms independently of the candidate (it has a
foreign key back to candidate). To make the application forms compatible with
the existing `CandidateUpserter` we need to establish a relationship between
the two models (we only pass the `Candidate` model to the upserter, so we need
to be able to navigate down to the forms from there).