Cache `Candidate` in Store if CRM integration is paused
When the CRM integration is paused, we want to provide a GUID to the Candidate, and cache it.

Usually, it is best practice to allow the CRM to generate sequential GUIDs which provide better SQL performance. However, in this scenario we have agreed it is beneficial to provide the GUID up-front because the School Experience app needs the Candidate ID immediately.

The Candidate upsert job is also triggered, so that when the CRM is operational again, we can persist the stored Candidate there, deleting the cache.