Update `UpsertCandidateJob` to remove candidate cache when successful
We cache the `Candidate` while the CRM integration is paused. However, we don't want to persist this permanently as it contains personal information. 

Ensure that when the `Candidate` is inserted into the CRM, the API cache record is deleted. Due to the way the `CandidateUpserter` job is configured, the API record will always be deleted within 24 hours (which it needs to be because it contains personal data).