Add `Candidate` cache
The CRM can't accept any new data during releases. The solution to this in Get into Teaching is to queue the request until the CRM is back up.

However, in School Experience, the Candidate upsert cannot be queued because the candidate information is needed immediately by the Schools Experience app.

Add a Candidate cache. The cache can be used when the CRM integration is paused.

Note: This creates a table for each entity related to Candidate, despite not declaring a DbSet for them.