Queue `CandidateUpsertJob` for sign up when CRM integration is paused
The CRM can't accept any new data during releases. The solution to this in Get into Teaching is to queue the request until the CRM is back up.

However, in School Experience, candidate information is needed immediately by the Schools Experience app. So, we must provide a GUID up-front to the Candidate before queuing the request. This way, the School Experience app can persist the CRM GUID when it receives a response (as it usually does when the CRM is operational).

Usually, it is best practice to allow the CRM to generate sequential GUIDs which provide better SQL performance. However, in this scenario we have agreed it is beneficial to provide the GUID up-front because the School Experience app needs the Candidate ID immediately.