Add an `AddClassroomExperienceNoteJob` so that request can be queued
Part of the School Experience candidate sign up (in the app) attempts to add a `ClassroomExperienceNote`. This firstly fetches the candidate from the CRM. 

When the CRM integration is paused, this is not possible and we were therefore seeing 404 not found.

To fix this, add an `AddClassroomExperienceNoteJob` so that the API can queue the request until the CRM is operational again. 

The `AddClassroomExperienceNoteJob` must complete after the `UpsertCandidateJob` of the `SignUp` action. However, the Hangfire jobs run concurrently, so we silently retry the `AddClassroomExperienceNoteJob` three times to allow some time for the `UpsertCandidateJob` to complete.