Sync new Apply API fields
The Apply team have added two additional fields to their application forms;
`submitted_at` and `recruitment_cycle_year`. We want to sync these through to
the CRM so that we can make the reporting consistent between Apply and the CRM.

Add new fields to the `ApplicationForm` models. Sync the new
`dfe_recruitmentyear` option set. Map fields as part of the
`FindApplyCandidateSyncJob`.