Throw exception when job fails
We wanted to silently retry the `AddClassroomExperienceNoteJob` three times before throwing an exception as the job may fail because a candidate has been added in another job.

Unfortunately, using the `Requeue` method results in the job ignoring the retry count, and being stuck in a loop between the `Enqueued` and `Processing` state.

Instead throw an exception (which causes Hangfire jobs to be rescheduled in an hours time), but add a bit more information in the message so that it is clear what is happening.