Dynamically map recruitment cycle year
Apply are sending through recruitment cycle years that are outwith the initial
option set values the CRM team created. The CRM team have since updated their
option set values, but we need to account for those in the API.

To make it easier to maintain going forward I'm mapping the recruitment cycle
year dynamically; there should never be a year before 2020 and the option set
values are guaranteed to be sequential.