Add field to track which welcome guide variant was served
If we serve a welcome guide to the user we want to be able to track which
version was served to them at the time of sign up (the fields to determine this
can change over time on the candidate record).

The field is not required and is left as plain text for the client to define.