Ensure correct method is called when the ID is API generated
Currently, the `CrmService` chooses the which `OrganisationService` method to call based on whether the model has an existing ID.

However, in certain circumstances, we need to allow to API to generate an ID up-front. 

Add a `GeneratedUpfrontId` method to the `BaseModel`, which sets a flag of `HasUpfrontId` on the model to true. The `BaseModel` can then choose which `OrganisationService` method to call based on the value of the flag.