Update `NewEntity` to handle when the model has an upfront ID
Currently, when an upfront ID is generated for a model, and it is saved in the CRM, it is returned with a different ID.

We may need to create the `Entity` object with an ID in the first place before adding it to the context.

Add an optional ID parameter that is passed down to the `OrganizationServiceAdapter`, we can pass this to the new `Entity` object.