Reduce event archive size to 4 months
Currently, if someone directly visits a past event on the website we allow them
to view it; whilst we want to retain this for the previous 4 months of events
that have an archive (Online Q&As) we don't want to let users view events that
are older than that. This change makes the API consistent with the website.