Relax event readable id validation
The `ReadableId` is used as the URL slug for an event; a number of pre-existing
events have invalid ids according to our existing regex used for validation.
This causes issues when these events are withdrawn/edited in the app. To avoid
replacing all of these slugs with valid ones and setting up redirects we're
instead going to relax the validation in the API - whilst retaining stricter
validation in the app - until the events are in the past. When they are, we
should be good to reinstate the stricter regex (the last event with an invalid
slug is currently 24/6/21).