Add a new claim to store the client name
We can currently only retrieve the token/role of the authenticated client,
however we need to be able to retrieve the name of the client so that we can
later use it when attributing TOTP actions that occur. This will enable us to
break down the TOTP actions by service (we can later make this more granular by
passing the transaction name from the clients).

Add a name claim when authenticating.