Populate ExistingCandidateRequest reference
As the clients do not yet send a reference explicitly we are going to default
them to the client name (which will give us a reasonable breakdown without
having to update each client). If we find this is not granular enough we can
update the clients to provide a reference relative to the individual
transactions.