Add new CandidatesSchoolExperience model
Currently, we add a ClassroomExperienceNote when the status of a School Experience candidate changes (e.g., they make a placement request, their placement is accepted etc.). This is a single formatted text field due to some time constraints during initial development.

The CRM team have created a new table - CandidateSchoolExperience - to represent this data. 

Add the initial models and validators. For now, the ClassroomExperienceNote will stay in place until the new feature is verified.