Get candidates with different primary and secondary phone numbers
Currently, we use secondary phone number in School Experience if primary number already has a value.  Going forward we have agreed to use primary phone number only. 

This endpoint will be used to retrieve all School Experience candidates with primary and secondary phone numbers that are different, because these will be negatively impacted by the change (we will be assuming that their primary phone number is the most up to date, but it will actually be their secondary phone number).

This endpoint will be removed once we have retrieved the candidates from production and passed them on to the CRM team.