Use other phone numbers when primary is null
We have simplified the School Experience sign up by removing alternate phone numbers. However, there are a number of records in the CRM that have School Experience (they have a Classroom Experience Note), but no primary phone number. 

If any of these candidates have open requests, they will not have a number in the Manage section of School Experience. Any new sign ups write to primary number so won't be a problem.

To fix this, attempt to fetch mobile, address, and secondary phone numbers from the Candidate model if primary number is null. The CRM team are going to fill the blanks, so this may be able to be removed shortly.